+++ resources/views/components/modal/variants/confirm.blade.php
@@ -1,6 +1,6 @@
-@props(['selectedItemId' => null, 'maxWidth' => null])
+@props(['selectedId' => null, 'maxWidth' => null])
 
-<x-gt-modal.base id="$selectedItemId" :$maxWidth {{ $attributes }}>
+<x-gt-modal.base id="$selectedId" :$maxWidth {{ $attributes }}>
     <div class="bx-title inline-flex va-c">
         <div class="pxy-05 bg-yellow-50 rounded-full">
             <x-gt-icon name="exclamation-triangle" class="txt-yellow" />
@@ -13,7 +13,7 @@
     </div>
 
     <div class="bx-footer tar">
-        <x-gt-button wire:click="$set('selectedItemId', false)" wire:loading.attr="disabled" text="Nevermind" />
+        <x-gt-button wire:click="$set('selectedId', false)" wire:loading.attr="disabled" text="Nevermind" />
         {{-- force the action to be passed through to make the modal more flexible --}}
         {{ $action }}
     </div>